Added my bidirectional RNN
Made a custom bidirectional RNN using the Torch library and got a decent score on the Kaggle competition (0.78486). 
I opted to not train a model using Torch's default RNN model because I wanted to learn more about the Torch library and the RNN architecture. I trained a model at the end to test my best model against the torch RNN model with the same parameters.